backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
cryptography==42.0.5
django-filter==23.5
cloudinary==1.39.1
django-cloudinary-storage==0.3.0

